Improve performance for add()
This optimization improves the performance when
adding an element to the set when you're already at
max_size (i.e we need to remove the lru element) by
inlining the removal of the head element from the linked
list.  Performance (using scripts/benchmark) shows roughly
an 8.8% improvement on my system with the
ten_times_max_size_sequential-lruset-LRUSet-10000 benchmark.